Move courseNumAsDecimal to analytics.ts
